+++ src/main/java/_1danhebojo/coalarm/coalarm_service/domain/user/repository/UserRepositoryImpl.java
@@ -1,8 +1,7 @@
 package _1danhebojo.coalarm.coalarm_service.domain.user.repository;
 
-import _1danhebojo.coalarm.coalarm_service.domain.user.entity.UserEntity;
+import _1danhebojo.coalarm.coalarm_service.domain.user.repository.entity.UserEntity;
 import _1danhebojo.coalarm.coalarm_service.domain.user.repository.jpa.UserJpaRepository;
-import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
